Work around a ridiculous Mac OS X-specific threading issue
I haven't reached the root of the issue but it looks like on at least OS X 10.6.x,
linking of CoreFoundation must happen on the main thread. If another thread happens to require 'thread'
before the main thread, it causes a crash.

There is surprisingly little information about this. The closest thing seems to be
http://bugs.python.org/issue7085 (see commit message of http://hg.python.org/cpython/rev/39989c66db17/).
Speaking of bros helping bros ;)