Remove 1.8.6 and 1.9.1 from ruby1, ruby2 and so on. Read on.
Most recent releases of both 1.8.6 and 1.9.1 fail to compile (1.9.1) or install gems (1.8.6). Both versions are
no longer in active use, are not actively maintained and majority of Ruby projects on travis-ci.org
does not use them.

It is a rough decision but travis-ci.org is maintained by literally a handful of people in their spare time.
Time to move on One True Ruby (1.8.6) and first rough cut at 1.9 (1.9.1).

I know, time flies by.