extract Manager::Factory and move dependencies further up the chain
@joshk, i recon this is a rather radical change again.

it follows the same pattern as with the Worker::Factory and the Build::Factory in travis-builds. you can see in the diff for the spec/manager_spec.rb how this significantly reduces the amount of stubbing necessary. 

i also see a great advantage in

1 - having all dependencies of the class "listed" in the top (i.e. you don't have to look through the implementation to find the dependencies. you just look at the initialize method and the attribute readers)
2 - having the actual assembling of all dependencies in one place (which is not deeply buried in a protected method nor clutters the class header)

i still think that factories are a rather clumpsy compromise here. ideally i'd only have a single factory (or module-level factory method). i did it this way for the following reasons:

1 - it seems to be good to have a factory method in places where there's a "cut". one cut is where travis-build is separated out. i felt another one is where the code looked like you might be using the Worker class in IRB for debugging/development? so i did not want to force you to pass all the dependencies manually and instead made another "cut" here
2 - we have lots of external dependencies which can behave quite nasty in tests (such as vbox, amqp, shell), i think it is good to have these created in one place so we can look at it and may be in a better position to decide where to finally move them
3 - i did not want to touch *all* of the code base

i am, all in all, happy with the way how the factories expose the weird dependency assembling we have. but i don't want to merge this in right away and thus push it to a branch and leave it up to you. also, i just don't know how you use these classes in IRB (or somewhere else), so feel free to change everything or just ignore this commit