Introduce Shell::Helpers#export_line
Right now we are trying to parse env variable pairs into key and value and
that is more challenging than it seems. We currently do not handle the following lines,
for example:

testcmd="ruby-test -r./test/helper.rb test/unit"

but we also don't have to. We can pass any sequence of key/value pairs to export
and it will export them all.

and try to deploy this fix quickly.