Milestone 1: spec suite passes, application boot part is done
 * Application connects, opens channels, declares queues and registers consumers
 * Test suite passes again, a few tests had to be fixed or updated
 * A couple irrelevant tests were removed

Next step is to go from using a single reporter throughout the entire worker
lifespan to instantiating a new one for each build. This will allow one worker
to build different types of projects (our short-term goal is PHP, Python, Perl)
as long as VM CI environment is capable of doing so.