Make each build queue consumer use a separate channel to work around some Heroku-specific oddities
On Heroku, RabbitMQ 2.5 ignores all consumers but the 1st one registered on a single channel.
This does not happen with 2.6 and 2.7 locally. I cannot tell why this is without bringing
this up to the RabbitMQ team so hence the workaround.

It has no downsides but technically should not be necessary. But we obey to Heroku gods.
For now.