Change the implementation of export_line to use FilteredString
Basically the idea is to be 'secure by default', which means that we should
display unfiltered string only in places there we're sure that it's ok (like
execute). All of the other places, where we want to just display it (like echo)
it will use to_s which displays filtered version.