change from using autoload in the worker to using declarative requires
this is a bit of a big commit, but this removes the use of autoload in the worker as well as moving the code in travis/worker to travis/worker/instance.

this means that Travis::Worker is more of a namespace than a place where code lives. also, it sets up the concept that Application is the code to start the worker, Pool looks after the pool of workers, and Instance is a single worker with uses a VM and Shell to run a job.