Print warning if we're unable to find/parse config
See travis-ci/travis-ci#814, and probably many more issues.

This is a very simple initial implementation for handling this problem.
We may want to actually put the parse error in the payload at some point
and print that, or maybe validate the .travis.yml file somehow, but this
works as a beginning.