Error a build if we couldn't parse the YAML file
Fixes travis-ci/travis-ci#814.

While there is an argument that no .travis.yml file should make the
build run using default settings for legacy reasons, if a user does
provide a .travis.yml file and it's not valid YAML then I can't think of
a reason why the user still would want the build to be run.