Filter Blue Box templates by 'pool'
Here, we introduce a new config parameter 'queue' in '.travis.yml'
which will optionally select which Blue Box template pool to select
the Blue Box template.

If there is no 'queue' in '.travis.yml', it behaves exactly the
same as before.

If there is, its value is used to match against the teamplates'
descriptions to find the templates which have the value (as a word)
in them.

For example, if we have a job with 'queue: update' in '.travis.yml',
we will choose a tamplate from the pool of templates that have the
descriptions of the form

  /^travis-update-([\w-]+)-\d{4}-\d{2}-\d{2}-\d{2}-\d{2}/

If no such template exists, we fall back to the current rule to
find
  /^travis-([\w-]+)-\d{4}-\d{2}-\d{2}-\d{2}-\d{2}/