Add second parameter in Blue Box image selection
Second parameter is now required to choose a non-production image
on Blue Box.

Following the suggestion made in https://github.com/travis-pro/ops/issues/326#issuecomment-52304661
we now look for images with the name of the form: travis-DIST-GROUP-TEMPLATE-%Y-%m-%d-%H-%M-SHA1
where `DIST` is specified in .travis.yml with the top-level `dist`,
`GROUP` with `group`, and `TEMPLATE` with `language`.